services/ticket-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
openai
langchain-openai
langchain-core
python-dotenv
bcrypt
python-jose[cryptography]
python-multipart
